Replace life curtain with simple cell count
In order to prove that a scene can run a simple simulation -- in this
case Conway's Game of Life -- adding a curtain containing a div for
every cell was the path of least resistance. This, however, proved to
create a serious limitation in rendering time. My hypothesis is that it
was the large overhead of manipulating DOM elements (over 600 of them)
that caused the slowdown, and not anything inherent to the simulation or
implementation of scenes and actors.

Half of that has already been borne out by removing the complex curtain
DOM elements and replacing it with a simple count of live cells. The
full simulation is still running and appears to be maintaining a very
high framerate, as we should expect.

The next test is to see whether a canvas-based visualization will be
more efficient, or whether rendering 600 to 1000 of anything is simply
prohibitive in the browser.