Add stage with Paper.js
Proof that my hypothesis was incorrect. It is indeed my naive
implementation of the Game of Life that is causing the slowdown. This is
probably good news. It means that drawing ~2000 shapes to the canvas
isn't inherently prohibitive.

As for the addition of the stage, this is a rudimentary implementation.
The stage is assigned right in the game js without leveraging anything
from src/stage.js. All it really does is initialize a set of symbol
instances for paper (one dot for each cell) and define an evaluate()
function that will loop through the dots and update their visibility to
match the associated cell's alive state.