Modify examples structure
In the midst of refactoring and updating documentation, it became apparent
that I didn't know how I expected the framework to be used. So, the example
of Conway's Game of Life became a major target of refactoring. Lots of
comments now explaining why and how it's put together the way it is.